Don't save empty descriptions
Prevent completely empty descriptions from being saved to the database.

Before this it would generate an empty description page prop and show a blank description on page info (common for File pages or if the page has no intro). A blank meta description isn't even added to the HTML, so it has no use.